[clang-c-frontend] refactored and added symbolic_types
